long long

+++ src/init.hpp
@@ -18,7 +18,7 @@ extern "C" void init(bool relations); // synonym ^^
 int share_memory();
 extern "C" void setMemoryLimit(long maxNodes0,long maxStatements0=-1,long maxChars0=-1);
 //char* share_memory(size_t key,long sizeOfSharedMemory,char* root,const void * desired);
-void* share_memory(size_t key, long sizeOfSharedMemory, void* root, const void * desired);
+void* share_memory(size_t key, long long sizeOfSharedMemory, void* root, const void * desired);
 void fixPointers();
 void fixPointers(Context* context);
 void fixNodeNames(Context* context, char* oldnodeNames);
